Move the Location generators to a separate trait
